Lockscreen Vibrate option 1 of 2
The new System setting Settings.System.LOCKSCREEN_VIBRATE_ENABLED
which defaults to 1, will effect the lockscreen on touch event
(when touching the unlock icon - or sliding it).

The GlowPadView (Slider) now has no Vibrate when the user selects the behaviour
in the Lockscreen Settings (see 2 of 2 patch).

No Restart is needed to make the change.

Change-Id: I59c52c6ea2f54427bbc883cd829952cefa58143c